requirements.txt
imapclient
python-dotenv
pyyaml
requests
beautifulsoup4
